requirements.txt
asgiref==3.8.1
audioread==3.0.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
contourpy==1.3.0
cycler==0.12.1
decorator==5.1.1
Django==5.1.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
fonttools==4.54.1
idna==3.10
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
matplotlib==3.9.2
msgpack==1.1.0
numba==0.60.0
numpy==2.0.2
packaging==24.1
pillow==11.0.0
platformdirs==4.3.6
pooch==1.8.2
pycparser==2.22
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
soundfile==0.12.1
soxr==0.5.0.post1
sqlparse==0.5.1
threadpoolctl==3.5.0
typing_extensions==4.12.2
urllib3==2.2.3
